# docs/test.py
# generated by datamodel-codegen:
#   filename:  openapi.yml

from __future__ import annotations

from datetime import datetime
from enum import Enum
from typing import List, Optional
from uuid import UUID

from pydantic import BaseModel, EmailStr, Field, SecretStr, constr


class SuccessResponseSchema(BaseModel):
    success: bool
    message: str
    error: Optional[str] = None


class ValidationErrorSchema(BaseModel):
    detail: Optional[List] = None


class Access(BaseModel):
    token: str
    expires_at: datetime


class Refresh(BaseModel):
    token: str
    expires_at: datetime


class TokensBaseSchema(BaseModel):
    access: Access
    refresh: Refresh


class TokenTypesEnum(Enum):
    access = 'access'
    refresh = 'refresh'


class AuthProvidersEnum(Enum):
    manual = 'manual'
    google = 'google'


class MailingMethodsEnum(Enum):
    email = 'email'
    sms = 'sms'


class AuthRequestSchema(BaseModel):
    email: EmailStr
    password: SecretStr


class AuthResponseSchema(BaseModel):
    email: EmailStr = Field(..., example='john@example.com')


class AuthVerifyRequestSchema(AuthRequestSchema):
    token: constr(min_length=6, max_length=6) = Field(..., example='123456')


class Token(BaseModel):
    token: str
    expires_at: datetime
    token_type: TokenTypesEnum


class RefreshTokenResponseSchema(SuccessResponseSchema):
    data: TokensBaseSchema


class UserBaseSchema(BaseModel):
    id: UUID
    email: EmailStr
    name: str
    username: str
    phone_number: str
    is_online: bool
    is_verified: bool
    auth_provider: AuthProvidersEnum
    mailing_method: MailingMethodsEnum
    created_at: datetime
    updated_at: Optional[datetime] = None


class Data(BaseModel):
    user: UserBaseSchema
    tokens: TokensBaseSchema


class AuthVerifyResponseSchema(SuccessResponseSchema):
    data: Data


class Data1(BaseModel):
    user: UserBaseSchema
    token: Token


class VerifyTokenResponseSchema(SuccessResponseSchema):
    data: Data1